clean and make runs displayed
